# Add performance benchmarks for CQC circuit routing