@tanujkhattar 2022-09-09
> The table has "failed" values for the largest circuit, please decrease the maximum size.

I've increased the timeout to 3 minutes for now. We can also have failed values for some combinations of a parameterized benchmark and the stored results and plots work as expected (no data points appear for a failed benchmark)

> The duration of asv run on my desktop is 45 minutes. Do we have some limit for the maximum duration of the benchmarks run?

This is expected for now since asv doesn't allow running benchmarks in parallel. Also, benchmarks are supposed to be executed on a cloud machine asynchronously and render plots of the form: https://cirq-infra.uc.r.appspot.com/